+++ api/src/main/java/dev/huskuraft/effortless/api/core/Player.java
@@ -64,5 +64,6 @@ public interface Player extends Entity {
 
     void resetStat(Stat<?> stat);
 
+    void causeFoodExhaustion(float amount);
 
 }